Update 'Day 2 - Diagonal_Difference' test code.

+++ hackerrank/1_Week_Prepare_Kit/Day_2/Diagonal_Difference/cpp11/run.bat
@@ -8,7 +8,7 @@ for %%i in (main*.in) do (
     echo @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
     echo --- input: %%i ---
     type %%i
-    del /q %%~ni.out
+    if exist %%~ni.out del /q %%~ni.out
     set OUTPUT_PATH=%%~ni.out
     build\Debug\main.exe < %%i
     echo +++ output: %%~ni.out +++
